stats-service/requirements.txt
clickhouse-driver
kafka-python
grpcio
grpcio-tools
protobuf
